meshtastic-discord-bridge-release/requirements.txt
# Core dependencies
meshtastic>=2.2.0
discord.py>=2.3.0
pyserial>=3.5
python-dotenv>=1.0.0

# Logging and utilities
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
